requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-telegram-bot>=20.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
openai>=0.27.0
pytest>=6.2.5
pytest-asyncio>=0.23.5
python-multipart>=0.0.5
pydantic>=1.8.2
tenacity>=8.0.1
tiktoken>=0.3.3
numpy>=1.21.2
pandas>=1.3.3
structlog>=24.1.0
python-json-logger>=2.0.7
asyncpg>=0.29.0
greenlet>=3.0.3
prometheus_client>=0.19.0

# CV Processing
python-docx>=1.1.2
pdfplumber>=0.11.5
